refactor(core): organize extras entities
