+++ lib/core/entities/movies/movie_detail.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
-import 'package:screen_pal/core/entities/genre.dart';
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 
 import 'movie_collection.dart';
