+++ lib/core/repo/extras_repo.dart
@@ -1,5 +1,5 @@
-import 'package:screen_pal/core/entities/genre.dart';
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 
 /// Extra details for the movies and tv_shows
 abstract interface class ExtrasRepo {
