+++ lib/core/use_cases/extras/get_languages.dart
@@ -1,4 +1,4 @@
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 import 'package:screen_pal/core/repo/extras_repo.dart';
 
 class GetLanguages {
