+++ lib/infrastructures/api/models/movies/raw_genre.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'package:screen_pal/core/entities/genre.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
 
 part 'raw_genre.freezed.dart';
 part 'raw_genre.g.dart';
