+++ lib/infrastructures/api/models/movies/raw_spoken_language.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 
 part 'raw_spoken_language.freezed.dart';
 part 'raw_spoken_language.g.dart';
