+++ lib/infrastructures/repo/extras_repo_impl.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:dio/dio.dart';
 
-import 'package:screen_pal/core/entities/genre.dart';
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 import 'package:screen_pal/core/repo/extras_repo.dart';
 import 'package:screen_pal/infrastructures/api/models/movies/raw_genre.dart';
 import 'package:screen_pal/infrastructures/api/models/movies/raw_spoken_language.dart';
