+++ lib/interfaces/providers/extras/genres_providers.dart
@@ -2,7 +2,7 @@
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import 'package:screen_pal/core/entities/genre.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
 import 'package:screen_pal/core/use_cases/extras/get_movie_genres.dart';
 import 'package:screen_pal/core/use_cases/extras/get_tv_show_genres.dart';
 import 'package:screen_pal/configs/service_locator/locator.dart' as service;
