+++ lib/interfaces/providers/extras/languages_provider.dart
@@ -2,7 +2,7 @@
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import 'package:screen_pal/core/entities/spoken_language.dart';
+import 'package:screen_pal/core/entities/extras/spoken_language.dart';
 import 'package:screen_pal/core/use_cases/extras/get_languages.dart';
 import 'package:screen_pal/configs/service_locator/locator.dart' as service;
