+++ test/helpers/mocks/repo/repositories.dart
@@ -2,10 +2,12 @@ import 'package:mockito/annotations.dart';
 import 'package:screen_pal/core/repo/app_settings_repo.dart';
 import 'package:screen_pal/core/repo/extras_repo.dart';
 import 'package:screen_pal/core/repo/movies_repo.dart';
+import 'package:screen_pal/core/repo/tv_shows_repo.dart';
 
 @GenerateMocks([
-  MoviesRepo,
-  ExtrasRepo,
   AppSettingsRepo,
+  ExtrasRepo,
+  MoviesRepo,
+  TvShowsRepo,
 ])
 void main() {}
