+++ test/helpers/providers/fake_movie_genres_notifier.dart
@@ -1,4 +1,4 @@
-import 'package:screen_pal/core/entities/genre.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
 import 'package:screen_pal/interfaces/providers/extras/genres_providers.dart';
 
 import '../dummy/dummy_genres.dart';
